Consolidate animation code
