+++ src/animations.h
@@ -1,0 +1,13 @@
+#ifndef ANIMATIONS_H
+#define ANIMATIONS_H
+
+#include "lvgl_wrapper.h"
+#include "config.h"
+
+// Forward declaration from the GigaAudio library
+class GigaAudio;
+
+void gauge_anim_cb(lv_timer_t *t);
+void voice_anim_cb(lv_timer_t *t);
+
+#endif
