--- src/gauge_anim.h
@@ -1,8 +1,0 @@
-#ifndef GAUGE_ANIM_H
-#define GAUGE_ANIM_H
-
-#include "lvgl_wrapper.h"
-
-void gauge_anim_cb(lv_timer_t *t);
-
-#endif
