--- src/voice_synth.h
@@ -1,14 +1,0 @@
-#ifndef VOICE_SYNTH_H
-#define VOICE_SYNTH_H
-
-#include "lvgl_wrapper.h"
-
-// Forward declaration from the GigaAudio library
-class GigaAudio;
-
-/// Timer callback used to drive the voice visualiser.
-/// The timer's user data must point to an ``GigaAudio`` instance so that
-/// the callback can query the current playback state.
-void voice_anim_cb(lv_timer_t *t);
-
-#endif
